Created utilities file and renamed Semaphore to stateMachines
