+++ Program/src/Application/stateMachines.cpp
@@ -1,0 +1,8 @@
+/*/*!
+ * @file Semaphore.cpp
+ * @par Author & Doxygen Editor
+ * 	Daniel Di Módica ~ <a href = "mailto: danifabriziodmodica@gmail.com">danifabriziodmodica@@gmail.com</a>
+ * @date 24/08/2023 01:02:47
+ */
+
+#include "statesMachines.h"
