+++ Program/src/Application/statesMachines.h
@@ -6,14 +6,12 @@
  * @brief Semaphore API (Application Programming Interface).
  */
 
-#ifndef SEMAPHORE_H_
-#define SEMAPHORE_H_
+#ifndef STATE_MACHINES_H_
+#define STATE_MACHINES_H_
 
 #include "ProgramConfig.h"
 #include "Timers.h"
 
-extern Timers g_timers_list;
 
-void delay(uint32_t milliseconds);
 
-#endif /* SEMAPHORE_H_ */
+#endif /* STATE_MACHINES_H_ */
