+++ Program/src/Hardware/06-MFRC522 RFID/MFRC522.h
@@ -12,7 +12,7 @@
 #include <cstring>
 #include "systick.h"
 #include "SPI.h"
-#include "Semaphore.h"
+#include "utilities.h"
 
 //////////////////////////////////////////
 // Page 0: Command and Status Registers //
@@ -185,9 +185,7 @@ private:
 	RFID_result_t PICC_REQA(uint8_t *bufferATQA, uint8_t *bufferSize);
 	virtual RFID_result_t PICC_Select(uint8_t validBits = 0);
 
-	// Support functions for debugging
-	char* strreverse(char* cstring);
-	char* byteToHEX(char* cstring, uint8_t value);
+
 public:
 	MFRC522(const Gpio& SCK, const Gpio& MOSI, const Gpio& MISO, const Gpio& SSEL, const Gpio& hardRST);
